pyproject.toml
[project]
name = "kobo-de-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dlt>=1.20.0",
    "duckdb>=1.4.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
]
